Fixed successful payment page being client-side and using metadata
